+++ ms-graph-api/src/MSGraphAPI/Users/User.hs
@@ -1,20 +1,26 @@
 -- | Users.User
-module MSGraphAPI.Users.User where
+module MSGraphAPI.Users.User (
+  get
+  , getMe
+  -- * types
+  , User(..)) where
 
 import GHC.Generics (Generic(..))
 
 -- aeson
-import qualified Data.Aeson as A (ToJSON(..), FromJSON(..), genericParseJSON, defaultOptions, Options(..), withObject, withText, (.:), (.:?), object, (.=))
+import qualified Data.Aeson as A (ToJSON(..), FromJSON(..), genericParseJSON)
 -- hoauth
 import Network.OAuth.OAuth2.Internal (AccessToken(..))
 -- req
 import Network.HTTP.Req (Req)
 -- text
-import Data.Text (Text, pack, unpack)
+import Data.Text (Text)
 
-import qualified MSGraphAPI.Internal.Common as MSG (get, post, aesonOptions)
+import qualified MSGraphAPI.Internal.Common as MSG (get, aesonOptions)
 
-
+-- | Representation of a user in the MS Graph API
+--
+-- https://learn.microsoft.com/en-us/graph/api/resources/users?view=graph-rest-1.0
 data User = User {
   uId :: Text
   , uUserPrincipalName :: Text
@@ -22,6 +28,7 @@ data User = User {
                  } deriving (Eq, Ord, Show, Generic)
 instance A.FromJSON User where
   parseJSON = A.genericParseJSON (MSG.aesonOptions "u")
+instance A.ToJSON User
 
 
 -- | Get user information